Move sub-classes of ToolStripSettings class to separate file
